refactor: extract letters array to separate module and import into script.js
