+++ index.html
@@ -8,6 +8,6 @@
     <link rel="stylesheet" href="style.css">
 </head>
 <body>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
